backend/requirements.txt
# pandas
# numpy
# dask[dataframe]
# langdetect
# requests
# click
# rich
# scrapy
# w3lib
# python-dateutil
# itemloaders
# python-dotenv
# snowflake-connector-python
# boto3
# gtts
# sodapy
# openai
# altair==5.3.0
# annotated-types==0.6.0
# anyio==4.3.0
# asn1crypto==1.5.1
# atpublic==4.1.0
# attrs==23.2.0
# blinker==1.7.0
# cachetools==5.3.3
# certifi==2024.2.2
# cffi==1.16.0
# chardet==5.2.0
# charset-normalizer==3.3.2
# click==8.1.7
# cloudpickle==2.2.1
# cryptography==42.0.5
# dnspython==2.6.1
# ecdsa==0.19.0
# exceptiongroup==1.2.0
# fastapi==0.110.1
# filelock==3.13.4
# fsspec==2024.3.1
# gitdb==4.0.11
# GitPython==3.1.43
# h11==0.14.0
# huggingface-hub==0.22.2
# idna==3.7
# Jinja2==3.1.3
# joblib==1.4.0
# jose==1.0.0
# jsonschema==4.21.1
# jsonschema-specifications==2023.12.1
# markdown-it-py==3.0.0
# MarkupSafe==2.1.5
# mdurl==0.1.2
# mpmath==1.3.0
# networkx
# numpy
# packaging
# pandas
# passlib
# pillow
# pinecone-client
# platformdirs
# protobuf
# pyarrow
# pyasn1==0.6.0
# pycparser==2.22
# pydantic==2.7.0
# pydantic_core==2.18.1
# pydeck==0.8.1b0
# Pygments==2.17.2
# PyJWT==2.8.0
# pymongo==4.6.3
# pyodbc==5.1.0
# pyOpenSSL==24.1.0
# python-dateutil==2.9.0.post0
# python-dotenv==1.0.1
# python-jose==3.3.0
# pytz==2024.1
# PyYAML==6.0.1
# referencing==0.34.0
# regex==2023.12.25
# reportlab==4.1.0
# requests==2.31.0
# rich==13.7.1
# rpds-py==0.18.0
# rsa==4.9
# safetensors==0.4.2
# scikit-learn==1.4.2
# scipy==1.13.0
# sentence-transformers==2.6.1
# six==1.16.0
# smmap==5.0.1
# sniffio==1.3.1
# snowflake==0.7.0
# snowflake-connector-python==3.8.1
# snowflake-snowpark-python==1.14.0
# snowflake._legacy==0.7.0
# snowflake.core==0.7.0
# sortedcontainers==2.4.0
# starlette==0.37.2
# streamlit==1.33.0
# streamlit-chat==0.1.1
# sympy==1.12
# tenacity==8.2.3
# threadpoolctl==3.4.0
# tokenizers==0.15.2
# toml==0.10.2
# tomlkit==0.12.4
# toolz==0.12.1
# torch==2.2.2
# tornado==6.4
# tqdm==4.66.2
# transformers==4.39.3
# typing_extensions==4.11.0
# tzdata==2024.1
# urllib3==1.26.18
# uvicorn==0.29.0
pandas
numpy
dask[dataframe]
langdetect
requests
click
rich
scrapy
w3lib
python-dateutil
itemloaders
python-dotenv
snowflake-connector-python
boto3
gtts
sodapy
openai
altair
annotated-types
anyio
asn1crypto
atpublic
attrs
blinker
cachetools
certifi
cffi
chardet
charset-normalizer
click
cloudpickle
cryptography
dnspython
ecdsa
exceptiongroup
fastapi
filelock
fsspec
gitdb
GitPython
h11
huggingface-hub
idna
Jinja2
joblib
jose
jsonschema
jsonschema-specifications
markdown-it-py
MarkupSafe
mdurl
mpmath
networkx
numpy
packaging
pandas
passlib
pillow
pinecone-client
platformdirs
protobuf
pyarrow
pyasn1
pycparser
pydantic
pydantic_core
pydeck
Pygments
PyJWT
pymongo
pyodbc
pyOpenSSL
python-dateutil
python-dotenv
python-jose
pytz
PyYAML
referencing
regex
reportlab
requests
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
sentence-transformers
six
smmap
sniffio
# snowflake
snowflake-connector-python
snowflake-snowpark-python
snowflake._legacy
snowflake.core
sortedcontainers
starlette
streamlit
streamlit-chat
sympy
tenacity
threadpoolctl
tokenizers
toml
tomlkit
toolz
torch
tornado
tqdm
transformers
typing_extensions
tzdata
urllib3
uvicorn
python-multipart